:art: Refactor: 패키지 변경
-주석 추가

Related to: #20
