+++ src/main/java/com/example/festimo/domain/admin/dto/AdminDTO.java
@@ -1,4 +1,4 @@
-package com.example.festimo.admin.dto;
+package com.example.festimo.domain.admin.dto;
 
 import java.time.LocalDateTime;
