+++ src/main/java/com/example/festimo/domain/admin/entity/Reviews.java
@@ -1,4 +1,4 @@
-package com.example.festimo.admin.entity;
+package com.example.festimo.domain.admin.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
