+++ src/main/java/com/example/festimo/domain/admin/mapper/AdminMapper.java
@@ -1,11 +1,11 @@
-package com.example.festimo.admin.mapper;
+package com.example.festimo.domain.admin.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
-import com.example.festimo.admin.entity.Users;
-import com.example.festimo.admin.dto.AdminDTO;
+import com.example.festimo.domain.admin.entity.Users;
+import com.example.festimo.domain.admin.dto.AdminDTO;
 
 @Mapper
 public interface AdminMapper {
