+++ src/main/java/com/example/festimo/domain/admin/repository/CommunityPostsRepository.java
@@ -1,7 +1,7 @@
-package com.example.festimo.admin.repository;
+package com.example.festimo.domain.admin.repository;
 
 
-import com.example.festimo.admin.entity.Post;
+import com.example.festimo.domain.admin.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
