+++ src/main/java/com/example/festimo/domain/admin/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.festimo.admin.repository;
+package com.example.festimo.domain.admin.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import lombok.NonNull;
 
-import com.example.festimo.admin.entity.Users;
+import com.example.festimo.domain.admin.entity.Users;
 
 public interface UserRepository extends JpaRepository<Users,Long> {
     @NonNull
